fix: pixels WASM + assert syntax.ol + stoll ambiguité GCC13
image_module: fallback GenImageColor(BLANK) si LoadImageFromTexture retourne
null sur WebGL (glReadPixels via FBO temp échoue sur une render texture
fraîchement créée jamais rendue)

parser: cast explicite static_cast<int64_t> sur stoll() pour lever
l'ambiguïté GCC13 entre NumberExpr(double) et NumberExpr(int64_t)
(long long != long sur Linux x86-64)

syntax.ol: corriger assert("x" == true) → assert(not not "x") —
== compare types et valeurs, pas la truthy-ness

+++ src/parser.cpp
@@ -776,7 +776,7 @@ std::unique_ptr<Expr> Parser::primary() {
     if (check(TokenType::NUMBER)) {
         const std::string& lex = advance().lexeme;
         if (lex.find('.') == std::string::npos)
-            return std::make_unique<NumberExpr>(std::stoll(lex));
+            return std::make_unique<NumberExpr>(static_cast<int64_t>(std::stoll(lex)));
         return std::make_unique<NumberExpr>(std::stod(lex));
     }
     if (check(TokenType::STRING))
